mv electron config to own file
